2.3.1 * email subject now changeable/customisable (see template wppizza-order-email-subject.php) * eliminate fatal error when plugin is re-installed without properly uninstalling it first (why would you do that anyway though ? oh well...) * BUGFIX when empty delete all categories was selected, only non empty categories were deleted * BUGFIX when using shortcode [wppizza category='non-existing-cat'] with a non-existing category, errors were thrown on page instead of just a blank page without menu items - 29th July 2013
git-svn-id: https://plugins.svn.wordpress.org/wppizza/trunk@747742 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/admin.empty.taxonomy.data.php
@@ -1,6 +1,6 @@
 <?php
+$terms = get_terms(''.$this->pluginSlugCategoryTaxonomy.'', array('hide_empty' => false));
 
-$terms = get_terms( $this->pluginSlugCategoryTaxonomy);
   if(count($terms)>0){
 	/*************************************************************************************************
 	*
@@ -13,15 +13,17 @@ $terms = get_terms( $this->pluginSlugCategoryTaxonomy);
 	if($the_query->have_posts()) {
 		$posts=$the_query->posts;	
 	}
+	if(isset($posts) && is_array($posts)){
 	foreach($posts as $k=>$v){
 		$postids[]=$v->ID;
-	}
+	}}
 	/*************************************************************************************************
 	*
 	*	[as attachments parents get set to 0 when a post is deleted , delete attachments first (ifset)]
 	*
 	*************************************************************************************************/
 	if($deleteAttachments){
+		if(isset($postids) && is_array($postids)){
 		foreach($postids as $k=>$v){
 			$args = array(
 			'post_parent' => $v,
@@ -32,16 +34,17 @@ $terms = get_terms( $this->pluginSlugCategoryTaxonomy);
 			foreach($attachments as $attachment){
 				wp_delete_attachment( $attachment->ID,true );
 			}
-		}
+		}}
 	}
 	/*************************************************************************************************
 	*
 	*	[now lets delete all posts]
 	*
 	*************************************************************************************************/
+	if(isset($postids) && is_array($postids)){
 	foreach($postids as $k=>$v){
 		wp_delete_post( $v, true );
-	}
+	}}
 	
 	/*************************************************************************************************
 	*
